Topbar, Schedule News bug fixed

+++ controller/adminController/newsController.js
@@ -1,5 +1,8 @@
 const db = require('../../models/index');
 const { sequelize } = require('../../models/index');
+const sharp = require('sharp');
+const fs = require('fs-extra');
+const moment = require('moment');
 
 let guid = () => {
   let s4 = () => {
@@ -19,42 +22,54 @@ const createNews = async (req, res, next) => {
   let id = guid();
   let img;
 
-  if (req.file) {
-    img = req.file.filename;
-  }
-  try {
-    await db.news.create({
-      id: id,
-      title: inputData.title,
-      key: inputData.key,
-      img,
-      categoryId: inputData.categoryId,
-      content: inputData.content,
-      status: inputData.status,
-      sharedAt: inputData.sharedAt,
-    });
+  console.log(inputData.sharedAt);
 
-    inputData.tags.split(",").forEach(async (tag) => {
-      await db.news_tags.create({
-        id: guid(),
-        newsId: id,
-        tagId: tag,
-      });
-    })
+  // try {
+  //   if(req.file) {
+  //     img = inputData.key + '.webp';
+  //     let imgPath = sharp(req.file.path)
+  //     .webp({ quality: 80 });
+  //     await imgPath.toFile('public/images/news/' + img);
+  //     fs.unlinkSync(req.file.path);
+  //   }
 
-    res.status(200).json({
-      status: 200,
-      statusText: 'Data has been successfully added to the database!',
-    });
-  } catch (error) {
-    return error;
-  };
+  //   await db.news.create({
+  //     id: id,
+  //     title: inputData.title,
+  //     key: inputData.key,
+  //     img,
+  //     categoryId: inputData.categoryId,
+  //     content: inputData.content,
+  //     status: inputData.status,
+  //     sharedAt: inputData.sharedAt,
+  //   });
 
+  //   inputData.tags.split(",").forEach(async (tag) => {
+  //     await db.news_tags.create({
+  //       id: guid(),
+  //       newsId: id,
+  //       tagId: tag,
+  //     });
+  //   })
+
+  //   res.status(200).json({
+  //     status: 200,
+  //     statusText: 'Data has been successfully added to the database!',
+  //   });
+  // } catch (error) {
+  //   res.status(500).json({
+  //     status: 500,
+  //     statusText: "Gözlənilməz xəta baş verdi. Xahiş olunur, daha sonra təkrar yoxlayasınız!",
+  //   });
+  //   if (req.file) {
+  //     fs.unlinkSync('public/images/news/' + img);
+  //   }
+  //   return error;
+  // };
 }
 
 const getSelectedNews = async (req, res, next) => {
   let id = req.query.id;
-  id = id.replace(':', '');
 
   try {
     let data = await db.news.findOne({
@@ -66,14 +81,19 @@ const getSelectedNews = async (req, res, next) => {
           model: sequelize.model('news_tags'),
           as: 'news_tags',
         }
-      ]
+      ],
     });
-    res.json(data);
+
+    // console.log(data);
+
+    data = JSON.stringify(data);
+
+    res.send(data);
   } catch (error) {
-    // res.status(500).json({
-    //   statusText: "Gözlənilməz xəta baş verdi. Xahiş olunur, daha sonra təkrar yoxlayasınız!",
-    //   error,
-    // });
+    res.status(500).json({
+      status: 500,
+      statusText: "Gözlənilməz xəta baş verdi. Xahiş olunur, daha sonra təkrar yoxlayasınız!",
+    });
     return error;
   }
 }
@@ -81,10 +101,10 @@ const getSelectedNews = async (req, res, next) => {
 const updateSelectedNews = async (req, res, next) => {
   let id = req.query.id;
   let inputData = req.body;
+  let img;
 
   try {
-
-    let news = await db.news.findOne({
+    let hasNews = await db.news.findOne({
       include: [
         {
           model: sequelize.model('news_tags'),
@@ -97,7 +117,7 @@ const updateSelectedNews = async (req, res, next) => {
       },
     });
 
-    const beforeTagIds = news.news_tags.map(tag => tag.tagId);
+    const beforeTagIds = hasNews.news_tags.map(tag => tag.tagId);
     const afterTagIds = inputData.tags.split(",");
     
     // Find IDs in afterIds that are not in beforeIds
@@ -106,14 +126,22 @@ const updateSelectedNews = async (req, res, next) => {
     // Find IDs in beforeIds that are not in afterIds
     const removeIds = beforeTagIds.filter(id => !afterTagIds.includes(id));
 
+    if (hasNews) {
+      if(req.file) {
+        img = inputData.key + '.webp';
+        let imgPath = sharp(req.file.path)
+        .webp({ quality: 80 });
+        await imgPath.toFile('public/images/news/' + img);
+        fs.unlinkSync(req.file.path);
+      }
 
-    await db.news.update({
-      title: inputData.title,
-      categoryId: inputData.categoryId,
-      content: inputData.content,
-      status: inputData.status,
-      sharedAt: inputData.sharedAt,
-    },
+      await db.news.update({
+        title: inputData.title,
+        categoryId: inputData.categoryId,
+        content: inputData.content,
+        status: inputData.status,
+        sharedAt: inputData.sharedAt,
+      },
       {
         where: {
           id,
@@ -122,46 +150,55 @@ const updateSelectedNews = async (req, res, next) => {
 
       if (req.file) {
         await db.news.update({
-          img: req.file.filename,
+          img,
         },
-          {
-            where: {
-              id,
-            }
-          })
-      }
-
-    if (addIds) {
-      addIds.forEach(async (tagId) => {
-        await db.news_tags.create({
-          id: guid(),
-          newsId: id,
-          tagId: tagId,
-        });
-      });
-    }
-
-    if (removeIds) {
-      removeIds.forEach(async (tagId) => {
-        await db.news_tags.destroy({
+        {
           where: {
-            newsId: id,
-            tagId,
+            id,
           }
+        })
+      }
+  
+      if (addIds) {
+        addIds.forEach(async (tagId) => {
+          await db.news_tags.create({
+            id: guid(),
+            newsId: id,
+            tagId: tagId,
+          });
         });
+      }
+  
+      if (removeIds) {
+        removeIds.forEach(async (tagId) => {
+          await db.news_tags.destroy({
+            where: {
+              newsId: id,
+              tagId,
+            }
+          });
+        });
+      }
+  
+      res.status(200).json({
+        status: 200,
+        statusText: "Məlumatlar uğurla yeniləndi!",
+      });
+    } else {
+      res.status(404).json({
+        status: 404,
+        statusText: "Belə bir xəbər tapılmadı!",
       });
     }
-
-    res.status(200).json({
-      status: 200,
-      statusText: "Məlumatlar uğurla yeniləndi!",
-    });
+    
   } catch (error) {
-    // res.status(500).json({
-    //   statusText: "Gözlənilməz xəta baş verdi. Xahiş olunur, daha sonra təkrar yoxlayasınız!",
-    //   error,
-    // });
-    console.log(error);
+    res.status(500).json({
+      status: 500,
+      statusText: "Gözlənilməz xəta baş verdi. Xahiş olunur, daha sonra təkrar yoxlayasınız!",
+    });
+    if (req.file) {
+      fs.unlinkSync('public/images/news/' + img);
+    }
     return error;
   }
 }
@@ -170,29 +207,75 @@ const deleteSelectedNews = async (req, res, next) => {
   let id = req.query.id;
 
   try {
-    await db.news.destroy({
+    let hasNews = await db.news.findOne({
       where: {
         id,
       }
-    })
-
-    await db.news_tags.destroy({
-      where: {
-        newsId: id,
-      }
-    })
-
-    res.status(200).json({
-      status: 200,
-      statusText: "Məlumatlar uğurla silindi!",
     });
+
+    if (hasNews) {
+      await db.news.destroy({
+        where: {
+          id,
+        }
+      });
+  
+      await db.news_tags.destroy({
+        where: {
+          newsId: id,
+        }
+      });
+
+      fs.unlinkSync('public/images/news/' + hasNews.img);
+      
+      res.status(200).json({
+        status: 200,
+        statusText: "Məlumatlar uğurla silindi!",
+      });
+    } else {
+      res.status(404).json({
+        status: 404,
+        statusText: "Belə bir xəbər tapılmadı!",
+      });
+    }
   } catch (error) {
-    // res.status(500).json({
-    //   statusText: "Gözlənilməz xəta baş verdi. Xahiş olunur, daha sonra təkrar yoxlayasınız!",
-    //   error,
-    // });
+    res.status(500).json({
+      status: 500,
+      statusText: "Gözlənilməz xəta baş verdi. Xahiş olunur, daha sonra təkrar yoxlayasınız!",
+    });
     return error;
   }
 }
 
-module.exports = { createNews, getSelectedNews, updateSelectedNews, deleteSelectedNews }
+const getAllNews = async (req, res, next) => {
+  try {
+    let allNews = await db.news.findAll({
+      include: [
+        {
+          model: sequelize.model('categories'),
+          as: 'category',
+        },
+        {
+          model: sequelize.model('news_tags'),
+          as: 'news_tag',
+          include: [
+            {
+              model: sequelize.model('tags'),
+              as: 'tag'
+            },
+          ],
+        },
+      ],
+    });
+
+    res.send(allNews);
+  } catch (error) {
+    res.status(500).send({
+      status: 500,
+      statusText: 'Gözlənilməz xəta baş verdi. Xahiş olunur, daha sonra təkrar yoxlayasınız!'
+    })
+    return error;
+  }
+}
+
+module.exports = { createNews, getSelectedNews, updateSelectedNews, deleteSelectedNews, getAllNews }
